+++ inst/examples/ex-independance.R
@@ -1,13 +1,11 @@
-## Abundance data
-## Coerce dataset to a count matrix (data from Desachy 2004)
-data("compiegne", package = "folio")
+data("cantabria")
 
 ## Compute EPPM
-counts_eppm <- eppm(compiegne)
+counts_eppm <- eppm(cantabria)
 plot_heatmap(counts_eppm) +
   khroma::scale_fill_YlOrBr(name = "EPPM")
 
 ## Compute PVI
-counts_pvi <- pvi(compiegne)
+counts_pvi <- pvi(cantabria)
 plot_heatmap(counts_pvi) +
   khroma::scale_fill_BuRd(name = "PVI", midpoint = 1)