+++ inst/examples/ex-plot_rank.R
@@ -1,7 +1,5 @@
-## Abundance matrix
-## Coerce datasets to a count matrix (data from Desachy 2004)
-data("compiegne", package = "folio")
+data("cantabria")
 
 ## Plot rank vs abundance
-plot_rank(compiegne)
-plot_rank(compiegne, facet = TRUE)
+plot_rank(cantabria)
+plot_rank(cantabria, facet = TRUE)